+++ src/components/MainContent/MainContent.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import CategoryFilter from "./CategoryFilter.jsx";
 import RestaurantList from "./RestaurantList.jsx";
-import restaurants from "../Data/Data.jsx";
+import restaurants from "../../Data/Data.jsx";
 
 function MainContent() {
   const [category, setCategory] = useState("전체");
